+++ modules/videoio/src/cap_avfoundation_mac.mm
@@ -104,7 +104,7 @@ public:
     bool grabFrame() CV_OVERRIDE;
     IplImage* retrieveFrame(int) CV_OVERRIDE;
     double getProperty(int property_id) const CV_OVERRIDE;
-    bool setProperty(int property_id, double value) CV_OVERRIDE;
+    bool setProperty(int property_id, int value) CV_OVERRIDE;
     int getCaptureDomain() /*const*/ CV_OVERRIDE { return cv::CAP_AVFOUNDATION; }
 
     virtual int didStart();
@@ -147,7 +147,7 @@ public:
     bool grabFrame() CV_OVERRIDE;
     IplImage* retrieveFrame(int) CV_OVERRIDE;
     double getProperty(int property_id) const CV_OVERRIDE;
-    bool setProperty(int property_id, double value) CV_OVERRIDE;
+    bool setProperty(int property_id, int value) CV_OVERRIDE;
     int getCaptureDomain() /*const*/ CV_OVERRIDE { return cv::CAP_AVFOUNDATION; }
 
     virtual int didStart();
@@ -507,7 +507,7 @@ double CvCaptureCAM::getProperty(int property_id) const{
     return retval;
 }
 
-bool CvCaptureCAM::setProperty(int property_id, double value) {
+bool CvCaptureCAM::setProperty(int property_id, int value) {
     NSAutoreleasePool *localpool = [[NSAutoreleasePool alloc] init];
 
     bool isSucceeded = false;
@@ -1102,7 +1102,7 @@ double CvCaptureFile::getProperty(int property_id) const{
     return 0;
 }
 
-bool CvCaptureFile::setProperty(int property_id, double value) {
+bool CvCaptureFile::setProperty(int property_id, int value) {
     if (mAsset == nil) return false;
 
     NSAutoreleasePool* localpool = [[NSAutoreleasePool alloc] init];
