+++ modules/videoio/src/cap_librealsense.hpp
@@ -19,7 +19,7 @@ public:
     virtual ~VideoCapture_LibRealsense();
 
     virtual double getProperty(int propIdx) const CV_OVERRIDE;
-    virtual bool setProperty(int propIdx, double propVal) CV_OVERRIDE;
+    virtual bool setProperty(int propIdx, int propVal) CV_OVERRIDE;
 
     virtual bool grabFrame() CV_OVERRIDE;
     virtual bool retrieveFrame(int outputType, OutputArray frame) CV_OVERRIDE;
