+++ modules/videoio/src/cap_winrt_capture.cpp
@@ -180,7 +180,7 @@ namespace cv {
     }
 
 
-    bool VideoCapture_WinRT::setProperty(int property_id, double value)
+    bool VideoCapture_WinRT::setProperty(int property_id, int value)
     {
         switch (property_id)
         {
